pyproject.toml
[tool.poetry]
name = "django_tutorial"
version = "0.0.1"
description = "Django Tutorial"
authors = ["José Ricardo Ciola Bricio <jr.ciola.bricio@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
pytest = "^7.1.2"
coverage = "^6.3.2"
httpx = "^0.22.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
